storage-infra/requirements.txt
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
mega.py==1.0.8
oauth2client==4.1.3
pathlib==1.0.1
pcloud==1.4
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.22.0
PyDrive==1.3.1
pyparsing==3.2.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
tenacity==5.1.5
uritemplate==4.1.1
urllib3==2.4.0
